remove an if
