+++ crates/burn-wgpu/src/compute/server.rs
@@ -186,7 +186,6 @@ where
         }
 
         let source = kernel.source().complete();
-        println!("{}", source);
         let pipeline = self.compile_source(&source);
         self.pipelines.insert(kernel_id.clone(), pipeline.clone());
