+++ crates/burn-wgpu/src/kernel/reduce/mod.rs
@@ -3,8 +3,8 @@ mod argmin_dim;
 mod base;
 mod mean_dim;
 mod naive_reduce_shader;
-mod reduction;
 mod shared_reduce_shader;
+mod sum;
 mod sum_dim;
 mod tune;
 
@@ -13,7 +13,7 @@ pub(crate) use argmin_dim::*;
 pub use base::*;
 pub(crate) use mean_dim::*;
 pub use naive_reduce_shader::*;
-pub use reduction::*;
 pub use shared_reduce_shader::*;
+pub use sum::*;
 pub(crate) use sum_dim::*;
 pub use tune::*;
